requirements.txt
django
django-cors-headers
django-plotly-dash
django_extensions
dash
pandas
pillow
codetiming
requests
numpy
PIL
qrcode
fpdf
pdfkit
google
goole-auth
mysqlcient=2.1.0

